+++ matlab/library/node_segmentation.cpp
@@ -13,7 +13,7 @@ void node_segmentation(std::vector<Mesh::Point>& points,
                       const PointSets& end_sets,
                       const std::vector<double>& voxel_dimensions,
                       const ShortestPathOptions& options,
-                      matrix<double>& visit_time,
+                      matrix<int>& visit_time,
                       matrix<int>& shortest_path_tree)
 {
   // Create functor handling regularization costs
